+++ molgenis-compute-core/src/main/java/org/molgenis/compute5/sysexecutor/AsyncStreamReader.java
@@ -1,5 +1,7 @@
 package org.molgenis.compute5.sysexecutor;
 
+import java.io.InputStream;
+
 /**
  * This class allows for the asynchronous reading of an input stream
  */
@@ -13,12 +15,13 @@ public interface AsyncStreamReader
 	public String getBuffer();
 
 	/**
-	 * 
+	 * Reads the {@link InputStream} used to instantiate this {@link AsyncStreamReader} and writes every line as an info
+	 * log
 	 */
 	public void run();
 
 	/**
-	 * 
+	 * When called stops the reading if supplied {@link InputStream}
 	 */
 	public void stopReading();
 }
